fix: re-enable erasable-syntax guards in catalog contents, lifting only Decorator
PR #5164 hot-fixed a catalog CI lint failure (the new Decorator selector
fired on @field/@tracked/@action in contents) by setting
'no-restricted-syntax': 'off' for all of contents/**. That also silently
un-banned enums, `import =`, `export =`, and runtime namespaces — which
are just as non-erasable and have no reason to be allowed in card modules.

Narrow the override to lift only the `Decorator` selector. Decorators are
permanently valid in contents (realm/Ember compilation pipeline; @field is
the card API), so this is the correct end state, not a grandfather — no
file-by-file exception list is needed (contents have zero violations of the
remaining selectors). The other four guards now apply to contents too,
matching the rest of the repo.

Extract the shared NO_COMPILATION_REQUIRED_TS_SELECTORS list (previously
duplicated verbatim in the root and ai-bot configs) into
eslint/erasable-syntax-selectors.cjs so the catalog override can filter it
without a third copy.
