+++ packages/catalog/.eslintrc.cjs
@@ -1,13 +1,28 @@
 'use strict';
 
+const {
+  NO_COMPILATION_REQUIRED_TS_SELECTORS,
+} = require('../../eslint/erasable-syntax-selectors.cjs');
+
+// contents/ files (both .ts card/command modules and .gts Glimmer components)
+// always go through the realm/Ember compilation pipeline, so decorators like
+// @field, @tracked, and @action are valid here — only the `Decorator` selector
+// is lifted. The remaining erasable-syntax guards (enum, `import =`,
+// `export =`, runtime namespaces) still apply, for consistency with the rest
+// of the repo.
+const CONTENTS_RESTRICTED_SYNTAX = [
+  'error',
+  ...NO_COMPILATION_REQUIRED_TS_SELECTORS.filter(
+    (s) => s.selector !== 'Decorator',
+  ),
+];
+
 module.exports = {
   overrides: [
     {
-      // contents/ .ts files are Boxel card/command modules that go through the
-      // realm compilation pipeline — decorators are valid here.
       files: ['contents/**/*.ts'],
       rules: {
-        'no-restricted-syntax': 'off',
+        'no-restricted-syntax': CONTENTS_RESTRICTED_SYNTAX,
       },
     },
     {
@@ -35,10 +50,7 @@ module.exports = {
         'plugin:prettier/recommended',
       ],
       rules: {
-        // .gts files always go through the Ember build pipeline, so decorators
-        // like @tracked and @action are valid here — the no-erasable-syntax
-        // restriction only applies to Node-native strip-types contexts.
-        'no-restricted-syntax': 'off',
+        'no-restricted-syntax': CONTENTS_RESTRICTED_SYNTAX,
         '@typescript-eslint/no-explicit-any': 'off',
         '@typescript-eslint/no-unused-vars': [
           'error',
